requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
langchain==0.1.3
langchain-openai==0.0.3
openai==1.9.0
python-dotenv==1.0.0
Werkzeug==2.3.5
